requirements.txt
asgiref
autopep8
Django
django-cleanup
django-froala-editor
Pillow
pycodestyle
sqlparse
toml
tzdata
